feat(config): move coonfig folder and fill doc
